+++ lib/ctx.h
@@ -388,7 +388,7 @@ extern uint32_t grn_gtick;
 #define GRN_CTX_TEMPORARY_DISABLE_II_RESOLVE_SEL_AND (0x40)
 
 typedef struct {
-  int32_t tv_sec;
+  int64_t tv_sec;
   int32_t tv_nsec;
 } grn_timeval;
